chore: clippy 清理 -- 0 warning + 代码审查
修复 18 个 clippy warning:
- from_str -> from_name (避免与 std::str::FromStr 冲突)
- link reference in list item (em-dash 分隔)
- clamp() 替换 min/max 模式
- sort_by_key 替换 sort_by
- iterator 替换 loop index
- 移除多余括号 / 未读赋值
- #[allow] 标注合理保留项 (too_many_arguments, dead_code embedder)

全量 486 测试通过, clippy 0 warning

+++ crates/shadow-providers/src/reliable.rs
@@ -81,7 +81,7 @@ impl RetryPolicy {
                 .map(|d| d.subsec_nanos() as u64)
                 .unwrap_or(0)
                 .wrapping_add(capped_ms.wrapping_mul(0x9E3779B97F4A7C15));
-            (seed % (jitter_range + 1)) as u64
+            seed % (jitter_range + 1)
         };
         let actual_ms = capped_ms + (capped_ms * jitter_factor / 100);
         Duration::from_millis(actual_ms.min(self.max_backoff_ms))
